TSemantic has been splitted in several files (and minor changes)
