+++ src/main/scala/sql-normal/tsemantic.scala
@@ -1,41 +1,6 @@
 package statelesser
 package sqlnormal
 
-import scalaz._, Leibniz.===
-
-sealed abstract class OpticKind
-case object KGetter extends OpticKind
-case object KAffineFold extends OpticKind
-case object KFold extends OpticKind
-
-case class TypeInfo(nme: TypeNme, isPrimitive: Boolean = false)
-
-case class OpticType[S, A](
-  kind: OpticKind, 
-  nme: Symbol, 
-  src: TypeInfo, 
-  tgt: TypeInfo)
-
-sealed abstract class TSel[S, A] {
-
-  def vars: Set[String] = this match {
-    case Pair(l, r, _) => l.vars ++ r.vars
-    case Just(e) => e.vars
-  }
-
-  def renameVars(rws: Set[(String, String)]): TSel[S, A] = this match {
-    case Pair(l, r, is) => Pair(l.renameVars(rws), r.renameVars(rws), is)
-    case Just(e) => Just(e.renameVars(rws))
-  }
-}
-
-case class Just[S, A](e: TExpr[S, A]) extends TSel[S, A]
-
-case class Pair[S, A, L, R](
-  l: TSel[S, L], 
-  r: TSel[S, R],
-  is: (L, R) === A) extends TSel[S, A]
-
 /*sealed*/ abstract class TSemantic[A]
 
 case class Done[O[_, _], S, A](
