+++ src/main/scala/sql/package.scala
@@ -7,7 +7,5 @@ package object `sql` {
   type Var = String
 
   type FieldName = String
-
-  object SQL extends FromSemantic with ToString
 }
